+++ sdl3-image-sys/ChangeLog.md
@@ -1,5 +1,8 @@
 # Changes
 
+- next
+    - Add metadata for structs and unions
+
 - 0.6.0:
     - Update sdl3-sys to 0.6.0
     - Synchronize version number with sdl3-sys